requirements.txt
Django
djangorestframework
django-cors-headers
djongo
slackclient
tweepy
